+++ CHANGE_LOG.md
@@ -6,8 +6,11 @@
 
 ### Components
 
+> **BREAKING CHANGE: Menu** - removed MenuItem `onClick` prop
+
 > **BREAKING CHANGE: ProgressionBoardThemeProvider** removed in favor of setting the theme at the global level
 
+- **Menu** - added `onSelect` prop
 - **DotDrawer** - added `ModalProps` optional prop, which serves as an object containing props applied to the Modal element
 - **ThemeProvider** - add `layer` object to global theme provider to match with UX mockups.
 - **ThemeProvider** - moved themes to global theme provider, `light`, `dark`, `agilityLight` and `agilityDark`